requirements.txt
click==8.1.3
cycler==0.11.0
Flask==2.1.2
Flask-WTF==1.0.1
fonttools==4.33.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.1.2
WTForms==3.0.1
